[TNT-258] refactor: 도메인 기반 구조에 맞게 프로필/회원 정보 수정 기능 리팩토링
