+++ src/main/java/com/tnt/member/dto/MemberInfo.java
@@ -8,7 +8,7 @@ import com.tnt.member.domain.SocialType;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 
-@Schema(description = "회원 조회 API 응답")
+@Schema(description = "회원 정보")
 public record MemberInfo(
 	@Schema(description = "회원 이름", example = "홍길동", nullable = false)
 	String name,
