+++ src/test/java/com/tnt/image/S3AdapterTest.java
@@ -16,6 +16,7 @@ import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.test.util.ReflectionTestUtils;
 
 import com.tnt.common.error.exception.ImageException;
+import com.tnt.image.infrastructure.S3Adapter;
 
 import software.amazon.awssdk.core.sync.RequestBody;
 import software.amazon.awssdk.services.s3.S3Client;
